backend-py/pyproject.toml
[project]
name = "sintra-trial-backend"
version = "1.0.0"
description = "Backend service for generating social media posts"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "openai>=1.103.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
]
